Introduce PhpMyAdmin compose bundle

+++ compose/docker-registry/compose.sh
@@ -2,4 +2,6 @@
 
 docker-compose $*
 
-sudo rm -f data/h2/prodDb.lock.db
+if [[ $(uname) == "Linux" ]]; then
+    sudo rm -f data/h2/prodDb.lock.db
+fi
